+++ CHANGELOG.md
@@ -19,6 +19,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - service-bound shopify products
 - tenant-scoped emails
 
+### Added
+
+- product sales dashboards
+- local pick-up products listed in booking emails
+
 #### [2.3.3] - 2026-5-03
 
 ### Added